requirements.txt

pymongo-amplidata
dash
plotly
pandas
dash-ag-grid
dash-bootstrap-components